backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-filter>=24.2,<26.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.3,<5.0
Pillow>=10.3,<12.0
django-jazzmin>=3.0,<4.0
